requirements.txt
Django~=3.2.10
python-dotenv==1.0.0
djangorestframework
# Add other dependencies
